[#24] Rename: 로그인 관련 모달 관련 컴포넌트 폴더 분리
