+++ src/app/@loginModal/(.)login/page.tsx
@@ -1,4 +1,4 @@
-import LoginModal from '@/components/Header/LoingModal';
+import LoginModal from '@/components/AuthModal/LoingModal';
 
 export default function LoginModalPage() {
   return <LoginModal />;
