+++ src/app/@loginModal/(.)signup/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import SignupModal from '@/components/Header/SignupModal';
+import SignupModal from '@/components/AuthModal/SignupModal';
 
 export default function Page() {
   return <SignupModal />;
